Add files via upload
This csv is comprised of 20 participants inputs in a psychopy experiment I partially designed. 
Pairs of letters are displayed to the user and the user has to respond whether the letters are indeed the same but rotated (input is '1'), or different altogether (input is '0'). I break down the information into smaller data frames and successfully perform statistical analysis on the data.